Subtle Refactors
