+++ app/src/androidTest/java/org/kiwix/kiwixmobile/nav/destination/library/LibraryRobot.kt
@@ -16,16 +16,15 @@
  *
  */
 
-package org.kiwix.kiwixmobile.nav.destination
+package org.kiwix.kiwixmobile.nav.destination.library
 
 import applyWithViewHierarchyPrinting
 import org.kiwix.kiwixmobile.BaseRobot
-import org.kiwix.kiwixmobile.Findable.Text
-import org.kiwix.kiwixmobile.Findable.ViewId
+import org.kiwix.kiwixmobile.Findable
 import org.kiwix.kiwixmobile.R
 import org.kiwix.kiwixmobile.language.LanguageRobot
 import org.kiwix.kiwixmobile.language.language
-import org.kiwix.kiwixmobile.testutils.TestUtils.getResourceString
+import org.kiwix.kiwixmobile.testutils.TestUtils
 
 /**
  * Authored by Ayush Shrivastava on 29/10/20
@@ -35,7 +34,7 @@ fun library(func: LibraryRobot.() -> Unit) = LibraryRobot().applyWithViewHierarc
 
 class LibraryRobot : BaseRobot() {
   init {
-    isVisible(ViewId(R.id.get_zim_nearby_device))
+    isVisible(Findable.ViewId(R.id.get_zim_nearby_device))
   }
 }
 
@@ -44,12 +43,12 @@ fun onlineLibrary(func: OnlineLibraryRobot.() -> Unit) =
 
 class OnlineLibraryRobot : BaseRobot() {
   init {
-    isVisible(Text(getResourceString(R.string.download)))
+    isVisible(Findable.Text(TestUtils.getResourceString(R.string.download)))
   }
 
   fun clickOnGlobeIcon(func: LanguageRobot.() -> Unit) {
-    clickOn(ViewId(R.id.select_language))
-    isVisible(Text(getResourceString(R.string.select_languages)))
+    clickOn(Findable.ViewId(R.id.select_language))
+    isVisible(Findable.Text(TestUtils.getResourceString(R.string.select_languages)))
     language(func)
   }
 }
